More complete (and ad-hoc tested) Django 1.7 support.
This is still not completely tested. I have a test project
that is not part of this project that I'm writing in 1.7,
and that uses this. Migrations work, but only the basic
migrations are guaranteed at this stage (create tables, add
index columns and FK relations).

The migrations testing is not included at this time.